requirements.txt
streamlit
pandas
numpy
pytesseract
Pillow
pdf2image
openai
python-dotenv